Implemented a new way of inserting the resources of ram memory and storage device, due to changing requirements in the system.

+++ src/main/java/com/pml/domain/StorageDevice.java
@@ -16,7 +16,7 @@ import com.pml.domain.enums.StorageDeviceType;
 public class StorageDevice extends Electronic {
 	private static final long serialVersionUID = 1L;
 	@NotNull
-	private Double sizeInGB;
+	private Integer sizeInGB;
 	@NotNull
 	private Integer architecture;
 	@NotNull
@@ -33,7 +33,7 @@ public class StorageDevice extends Electronic {
 	}
 	
 	public StorageDevice(Long id, Date createdDate, Date lastModifiedDate, String manufacturer,
-			String model, String description, boolean itWorks, Double sizeInGB, 
+			String model, String description, boolean itWorks, Integer sizeInGB, 
 			StorageDeviceArchitecture storageDeviceArchitecture, StorageDeviceType type, Computer computer) {
 		super(id, createdDate, lastModifiedDate, EquipmentType.STORAGE_DEVICE, manufacturer, model, description, itWorks, false);
 		this.sizeInGB = sizeInGB;
@@ -43,11 +43,11 @@ public class StorageDevice extends Electronic {
 		this.setItComposed(false);
 	}
 		
-	public Double getSizeInGB() {
+	public Integer getSizeInGB() {
 		return sizeInGB;
 	}
 	
-	public void setSizeInGB(Double sizeInGB) {
+	public void setSizeInGB(Integer sizeInGB) {
 		this.sizeInGB = sizeInGB;
 	}
