+++ src/main/java/com/pml/dto/ComputerDTO.java
@@ -31,8 +31,8 @@ public class ComputerDTO implements Serializable {
 	private Integer operatingSystem;
 	private Integer operatingSystemArchitecture;
 	private boolean onTheDomain;
-	private Double totalRamMemory;
-	private Double totalStorageMemory;
+	private Integer totalRamMemory;
+	private Integer totalStorageMemory;
 	
 	public ComputerDTO() {
 	}
@@ -207,19 +207,19 @@ public class ComputerDTO implements Serializable {
 		this.onTheDomain = onTheDomain;
 	}
 
-	public Double getTotalRamMemory() {
+	public Integer getTotalRamMemory() {
 		return totalRamMemory;
 	}
 
-	public void setTotalRamMemory(Double totalRamMemory) {
+	public void setTotalRamMemory(Integer totalRamMemory) {
 		this.totalRamMemory = totalRamMemory;
 	}
 	
-	public Double getTotalStorageMemory() {
+	public Integer getTotalStorageMemory() {
 		return totalStorageMemory;
 	}
 
-	public void setTotalStorageMemory(Double totalStorageMemory) {
+	public void setTotalStorageMemory(Integer totalStorageMemory) {
 		this.totalStorageMemory = totalStorageMemory;
 	}
