+++ src/main/java/com/pml/dto/StorageDeviceDTO.java
@@ -22,7 +22,7 @@ public class StorageDeviceDTO implements Serializable {
 	private String description;
 	private boolean itWorks; 
 	private boolean itComposed;
-	private Double sizeInGB;
+	private Integer sizeInGB;
 	private Integer architecture;
 	private Integer type;
 	
@@ -126,11 +126,11 @@ public class StorageDeviceDTO implements Serializable {
 		this.itComposed = isItComposed;
 	}
 
-	public Double getSizeInGB() {
+	public Integer getSizeInGB() {
 		return sizeInGB;
 	}
 
-	public void setSizeInGB(Double sizeInGB) {
+	public void setSizeInGB(Integer sizeInGB) {
 		this.sizeInGB = sizeInGB;
 	}
