+++ src/main/java/com/pml/dto/StorageDeviceNewDTO.java
@@ -16,7 +16,7 @@ public class StorageDeviceNewDTO implements Serializable {
 	private String description;
 	private boolean itWorks = true;
 	@NotNull(message = "This field is mandatory")
-	private Double sizeInGB;
+	private Integer sizeInGB;
 	@NotNull(message = "This field is mandatory")
 	private Integer architecture;
 	@NotNull(message = "This field is mandatory")
@@ -58,11 +58,11 @@ public class StorageDeviceNewDTO implements Serializable {
 		this.itWorks = itWorks;
 	}
 	
-	public Double getSizeInGB() {
+	public Integer getSizeInGB() {
 		return sizeInGB;
 	}
 
-	public void setSizeInGB(Double sizeInGB) {
+	public void setSizeInGB(Integer sizeInGB) {
 		this.sizeInGB = sizeInGB;
 	}
